[#112] ✨ refactor: 오타 수정 및 카트/상품/상품 이미지 조인 로직 추가
